Add auth_ldap2 module
This module enables non-basic default LDAP instances to be used.   The
original module was incapble of being used against any actual trees as
well as with any DN's that didn't follow a specific pattern.

This module gives users of ldap more control over the filters and
attributes that matter for their ldap service.